Rename PasswordResetException to PasswordResetBaseException for integrity
